Test autonomous system API exposure.
